requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
filelock==3.16.1
fonttools==4.55.8
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.19.3
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.3.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
py-rouge==1.1
pydantic==2.10.5
pydantic_core==2.27.2
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge==1.0.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.5.1
tokenizers==0.15.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.38.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.5
urllib3==1.26.6
